requirements.txt
-i https://pypi.org/simple
absl-py==0.10.0
astor==0.8.1
attrs==20.2.0
blis==0.2.4
boto3==1.10.19
botocore==1.13.19
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cycler==0.10.0
cymem==2.0.3
cython==0.29.21
dill==0.3.2
docutils==0.15.2
dotmap==1.3.17
fairseq==0.9.0
flask-cors==3.0.9
flask==1.1.2
ftfy==5.7
future==0.18.2
gast==0.2.2
gitdb==4.0.5
gitpython==3.1.0
google-auth-oauthlib==0.4.1
google-auth==1.21.2
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
h5py==2.10.0
idna==2.8
importlib-metadata==2.0.0 ; python_version < '3.8'
itsdangerous==1.1.0
jieba==0.42.1
jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
keras-applications==1.0.8
keras-preprocessing==1.1.2
keras==2.4.3
kiwisolver==1.2.0
markdown==3.2.2
markupsafe==1.1.1
matplotlib==3.1.2
mock==4.0.1
more-itertools==8.5.0
murmurhash==1.0.2
nltk==3.4.5
numpy==1.17.4
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
plac==0.9.6
pluggy==0.13.1
portalocker==2.0.0
preshed==2.0.1
promise==2.3
protobuf==3.13.0
psutil==5.6.5
ptvsd==4.3.2
py==1.9.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20
pyparsing==2.4.7
pytest==5.3.5
pythainlp==2.1.4
python-dateutil==2.8.0 ; python_version >= '2.7'
pyyaml==5.3.1
regex==2019.11.1
requests-oauthlib==1.3.0
requests==2.22.0
rsa==4.6 ; python_version >= '3.5'
s3transfer==0.2.1
sacrebleu==1.4.14
sacremoses==0.0.35
scikit-learn==0.22.1
scipy==1.4.1
sentencepiece==0.1.83
seqeval==0.0.12
six==1.14.0
smmap==3.0.4
spacy==2.1.9
srsly==1.0.2
tensorboard==2.1.1
tensorboardx==2.0
tensorflow-datasets==2.1.0
tensorflow-estimator==2.1.0
tensorflow-metadata==0.24.0
tensorflow==2.1.0
termcolor==1.1.0
thinc==7.0.8
tinydb==4.1.1
torch==1.6.0
tqdm==4.38.0
urllib3==1.25.10 ; python_version >= '3.4'
wasabi==0.8.0
wcwidth==0.2.5
werkzeug==1.0.1
wheel==0.35.1 ; python_version >= '3'
wrapt==1.12.1
zipp==3.2.0
